requirements/base.txt
kubernetes==6.0.0
aiohttp==3.2.1
python-decouple==3.1
attrdict==2.0.0